requirements.txt
agency-swarm
docker
pydantic
PyYAML
requests
python-dotenv
pandas
